Add basic interaction with DB object
In the future, the idea of the DB object will probably change to become something more generic, but for the moment I'm using it as a seam to separate the tracker logic from its persistence layer.

This commit introduces some basic checking for whether a DB file exits. All of these calls are being tested by using a FakeDb mock object and testing calls to that mock object. The basic functionality in place (in the form of API calls to the DB) is checking if the DB already exists, creating a new DB, and adding a new habit to the DB.

I'm using a slightly experimental technique for stubbing methods on the FakeDb class by importing the `New Method` library and using that functionality to dynamically add methods to my FakeDb class. When asserting that a message was sent to the mock object, I'm just setting a `method_called` boolean property on the mock object to `True` if the message was sent, and basing my assertion for the test on that.